refactor(AlertDialog): refactor `AlertDialog` components
