+++ src/components/alert-dialog/alert-dialog-header.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 
-import { twMerge } from "tailwind-merge";
+import { cn } from "~/utils";
 
 export type AlertDialogHeaderProps = React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>;
 
 export function AlertDialogHeader({ className, ...props }: AlertDialogHeaderProps) {
-  return <div className={twMerge("mb-4 flex flex-col space-y-2 text-center sm:text-left", className)} {...props} />;
+  return <div className={cn("mb-4 flex flex-col space-y-2 text-center sm:text-left", className)} {...props} />;
 }
